# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: Slides API
  version: "2"
  description: |
    Sync-and-play slideshow API. The server is the source of truth for **configuration
    rules** (`AppSettings` grouped by `query`, `playback`, `configuration`) and a **real-time
    broker** (SSE) for domain-scoped settings updates. The server does not track playback
    state (timers, indices).

    **Kiosk boot:** `GET /meta` → `GET /settings` + `GET /events` (SSE) in parallel →
    `POST /slideshow/query` with query fields from settings → local timer and index.

    **Remote control:** `PATCH /settings/{domain}` (e.g. Home Assistant) broadcasts
    domain events on the SSE stream.
servers:
  - url: /api/v1
    description: Versioned domain API (relative to the Slides host)
paths:
...
  /settings:
    get:
      operationId: getSettings
      summary: Effective configuration
      description: |
        Returns the fully effective configuration: `DEFAULT_*` env defaults
        shallow-merged with persisted per-domain overrides (`settings.query.json`,
        `settings.playback.json`, `settings.configuration.json`), then optional URL query
        overrides using bracket notation (e.g. `?query[shuffle]=false&playback[intervalMs]=5000`;
        array values use comma separation `query[albumIds]=id1,id2`). URL overrides are
        session-only and are not persisted. Ignored params: `seed`, `cursor`, `limit`.
      parameters:
        - name: query
          in: query
          required: false
          style: deepObject
          explode: true
          schema:
            type: object
          description: Example — domain keys may be passed via bracket notation.
      responses:
        '200':
          description: Effective settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppSettings'
        '500':
          $ref: '#/components/responses/ServerError'

    delete:
      operationId: deleteSettings
      summary: Clear all persisted overrides
      description: |
        Removes all persisted overrides (all three domain files), reverting to env defaults.
        Broadcasts `settings_cleared` on SSE.
      responses:
        '200':
          description: Overrides cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Settings overrides cleared
  /settings/query:
    patch:
      operationId: patchQuerySettings
      summary: Partially update query settings overrides
      description: |
        Shallow-merges the partial `QuerySettingsUpdate` body with existing persisted
        query overrides in `settings.query.json` and returns the full effective
        `AppSettings`. Broadcasts `query_updated` on SSE.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuerySettingsUpdate'
      responses:
        '200':
          description: Effective settings after update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppSettings'
        '400':
          $ref: '#/components/responses/ClientError'

    delete:
      operationId: deleteQuerySettings
      summary: Clear query settings overrides
      description: |
        Removes query overrides, reverting query domain to env defaults.
        Broadcasts `query_updated` on SSE.
      responses:
        '200':
          description: Query overrides cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Query settings overrides cleared
  /settings/playback:
    patch:
      operationId: patchPlaybackSettings
      summary: Partially update playback settings overrides
      description: |
        Shallow-merges the partial `PlaybackSettingsUpdate` body with existing persisted
        playback overrides in `settings.playback.json` and returns the full effective
        `AppSettings`. Broadcasts `playback_updated` on SSE.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaybackSettingsUpdate'
      responses:
        '200':
          description: Effective settings after update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppSettings'
        '400':
          $ref: '#/components/responses/ClientError'

    delete:
      operationId: deletePlaybackSettings
      summary: Clear playback settings overrides
      description: |
        Removes playback overrides, reverting playback domain to env defaults.
        Broadcasts `playback_updated` on SSE.
      responses:
        '200':
          description: Playback overrides cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Playback settings overrides cleared
  /settings/configuration:
    patch:
      operationId: patchConfigurationSettings
      summary: Partially update configuration settings overrides
      description: |
        Shallow-merges the partial `ConfigurationSettingsUpdate` body with existing persisted
        configuration overrides in `settings.configuration.json` and returns the full effective
        `AppSettings`. Broadcasts `configuration_updated` on SSE.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigurationSettingsUpdate'
      responses:
        '200':
          description: Effective settings after update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppSettings'
        '400':
          $ref: '#/components/responses/ClientError'

    delete:
      operationId: deleteConfigurationSettings
      summary: Clear configuration settings overrides
      description: |
        Removes configuration overrides, reverting configuration domain to env defaults.
        Broadcasts `configuration_updated` on SSE.
      responses:
        '200':
          description: Configuration overrides cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Configuration settings overrides cleared
...
components:
...
  schemas:
...
    AppSettings:
      type: object
      required: [query, playback, configuration]
      properties:
        query:
          $ref: '#/components/schemas/QuerySettings'
        playback:
          $ref: '#/components/schemas/PlaybackSettings'
        configuration:
          $ref: '#/components/schemas/ConfigurationSettings'
    QuerySettings:
      type: object
      required:
        - albumIds
        - albumOperator
        - personIds
        - personOperator
        - excludeAlbumIds
        - excludePersonIds
        - globalOperator
        - shuffle
      properties:
        albumIds:
          type: array
          items: { type: string }
        albumOperator:
          $ref: '#/components/schemas/FilterOperator'
        personIds:
          type: array
          items: { type: string }
        personOperator:
          $ref: '#/components/schemas/FilterOperator'
        excludeAlbumIds:
          type: array
          items: { type: string }
        excludePersonIds:
          type: array
          items: { type: string }
        locationCountry:
          type: string
        locationState:
          type: string
        locationCity:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        globalOperator:
          $ref: '#/components/schemas/FilterOperator'
        shuffle:
          type: boolean
    PlaybackSettings:
      type: object
      required:
        - intervalMs
        - autoplay
        - layout
        - photoScaleMode
        - transitionType
        - transitionDuration
        - photoAnimationType
        - photoAnimationDuration
        - photoAnimationIntensity
        - livePhotoEnabled
        - livePhotoDelay
      properties:
        intervalMs:
          type: integer
        autoplay:
          type: boolean
        layout:
          type: string
          enum: [single, split]
        photoScaleMode:
          type: string
          enum: [fit_inside, fill_crop, stretch, original]
        transitionType:
          type: string
          enum: [fade, slide, none]
        transitionDuration:
          type: number
        photoAnimationType:
          type: string
          enum: [none, zoom-in, zoom-out, pan, ken-burns]
        photoAnimationDuration:
          type: number
        photoAnimationIntensity:
          type: number
        livePhotoEnabled:
          type: boolean
        livePhotoDelay:
          type: number
    ConfigurationSettings:
      type: object
      required:
        - photoMetadataDateFormat
        - clockUse24HourFormat
        - clockDateFormat
        - weatherLat
        - weatherLng
      properties:
        photoMetadataDateFormat:
          type: string
        clockUse24HourFormat:
          type: boolean
        clockDateFormat:
          type: string
        weatherLat:
          type: number
          format: double
        weatherLng:
          type: number
          format: double
    QuerySettingsUpdate:
      type: object
      properties:
        albumIds:
          type: array
          items: { type: string }
        albumOperator:
          $ref: '#/components/schemas/FilterOperator'
        personIds:
          type: array
          items: { type: string }
        personOperator:
          $ref: '#/components/schemas/FilterOperator'
        excludeAlbumIds:
          type: array
          items: { type: string }
        excludePersonIds:
          type: array
          items: { type: string }
        locationCountry:
          type: string
        locationState:
          type: string
        locationCity:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        globalOperator:
          $ref: '#/components/schemas/FilterOperator'
        shuffle:
          type: boolean
    PlaybackSettingsUpdate:
      type: object
      properties:
        intervalMs:
          type: integer
        autoplay:
          type: boolean
        layout:
          type: string
          enum: [single, split]
        photoScaleMode:
          type: string
          enum: [fit_inside, fill_crop, stretch, original]
        transitionType:
          type: string
          enum: [fade, slide, none]
        transitionDuration:
          type: number
        photoAnimationType:
          type: string
          enum: [none, zoom-in, zoom-out, pan, ken-burns]
        photoAnimationDuration:
          type: number
        photoAnimationIntensity:
          type: number
        livePhotoEnabled:
          type: boolean
        livePhotoDelay:
          type: number
    ConfigurationSettingsUpdate:
      type: object
      properties:
        photoMetadataDateFormat:
          type: string
        clockUse24HourFormat:
          type: boolean
        clockDateFormat:
          type: string
        weatherLat:
          type: number
          format: double
        weatherLng:
          type: number
          format: double
    FilterOperator:
      type: string
      enum: [AND, OR]
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message]
          properties:
            type:
              type: string
              enum: [network, auth, client, server]
            message:
              type: string
            code:
              type: string
  responses:
    ClientError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'